HomeWork 8. rework_3
