+++ index.html
@@ -36,12 +36,7 @@
     </div>
 
     <div class="control">
-      <!--            <button class="button" id="btn-kick">-->
-      <!--                Thunder Jolt-->
-      <!--            </button>-->
-      <!--            <button class="button" id="btn-kick-spec">-->
-      <!--                Fireball-->
-      <!--            </button>-->
+
     </div>
 
     <div class="pokemon player2">
