+++ utils.js
@@ -1,6 +1,3 @@
-// random = (num) => Math.ceil(Math.random() * num)
-
-// function random (num)  {Math.ceil(Math.random() * num)};
 function random(num) {
     return Math.ceil(Math.random() * num);
 }
